Removed slower functions.
For the purpose of benchmarking, I copied the following
Hamming function:
https://github.com/niven/simhashing/blob/master/util.go#L48
which is just Kernighan's hamming function implemented:
http://graphics.stanford.edu/~seander/bithacks.html#CountBitsSetKernighan

This hamming function is slower than the DistanceBits
implementation, which is slower than a lookup table:
BenchmarkDistance         2000      845001 ns/op
BenchmarkHamDistance       500     3398084 ns/op
BenchmarkDistanceBits      500     3005490 ns/op

Both DistanceBits and HammingDistance functions are removed with this commit.

All benchmarks are removed and the test (but the function is correct).